+++ ios/FastImage/FFFastImageViewManager.h
@@ -1,5 +1,9 @@
+#if !defined(RCT_NEW_ARCH_ENABLED) || RCT_NEW_ARCH_ENABLED == 0
+
 #import <React/RCTViewManager.h>
 
 @interface FFFastImageViewManager : RCTViewManager
 
 @end
+
+#endif