+++ ios/FastImage/FFFastImageViewManager.mm
@@ -1,11 +1,8 @@
+#if !defined(RCT_NEW_ARCH_ENABLED) || RCT_NEW_ARCH_ENABLED == 0
+
 #import "FFFastImageViewManager.h"
 #import "FFFastImageView.h"
 
-#import <SDWebImage/SDImageCache.h>
-#import <SDWebImage/SDWebImagePrefetcher.h>
-#ifdef RCT_NEW_ARCH_ENABLED
-#import <RNFastImageSpec/RNFastImageSpec.h>
-#endif
 @implementation FFFastImageViewManager
 
 RCT_EXPORT_MODULE(FastImageView)
@@ -25,38 +22,6 @@ RCT_EXPORT_VIEW_PROPERTY(onFastImageLoadEnd, RCTDirectEventBlock)
 RCT_REMAP_VIEW_PROPERTY(tintColor, imageColor, UIColor)
 RCT_EXPORT_VIEW_PROPERTY(blurRadius, CGFloat)
 
-RCT_EXPORT_METHOD(preload:(nonnull NSArray<FFFastImageSource *> *)sources)
-{
-    NSMutableArray *urls = [NSMutableArray arrayWithCapacity:sources.count];
-
-    [sources enumerateObjectsUsingBlock:^(FFFastImageSource * _Nonnull source, NSUInteger idx, BOOL * _Nonnull stop) {
-        [source.headers enumerateKeysAndObjectsUsingBlock:^(NSString *key, NSString* header, BOOL *stop) {
-            [[SDWebImageDownloader sharedDownloader] setValue:header forHTTPHeaderField:key];
-        }];
-        [urls setObject:source.url atIndexedSubscript:idx];
-    }];
-
-    [[SDWebImagePrefetcher sharedImagePrefetcher] prefetchURLs:urls];
-}
-
-RCT_EXPORT_METHOD(clearMemoryCache:(RCTPromiseResolveBlock)resolve reject:(RCTPromiseRejectBlock)reject)
-{
-    [SDImageCache.sharedImageCache clearMemory];
-    resolve(NULL);
-}
-
-RCT_EXPORT_METHOD(clearDiskCache:(RCTPromiseResolveBlock)resolve reject:(RCTPromiseRejectBlock)reject)
-{
-    [SDImageCache.sharedImageCache clearDiskOnCompletion:^(){
-        resolve(NULL);
-    }];
-}
-#ifdef RCT_NEW_ARCH_ENABLED
-- (std::shared_ptr<facebook::react::TurboModule>)getTurboModule:
-    (const facebook::react::ObjCTurboModule::InitParams &)params
-{
-    return std::make_shared<facebook::react::NativeFastImageViewSpecJSI>(params);
-}
-#endif
-
 @end
+
+#endif