+++ src/index.tsx
@@ -21,8 +21,8 @@ import {
 const isFabricEnabled = (global as any)?.nativeFabricUIManager != null
 const isTurboModuleEnabled = (global as any).__turboModuleProxy != null
 const FastImageViewModule = isTurboModuleEnabled
-    ? require('./NativeFastImageView').default
-    : NativeModules.FastImageView
+    ? require('./NativeFastImageViewModule').default
+    : NativeModules.FastImageViewModule
 
 const FastImageView = isFabricEnabled
     ? require('./FastImageViewNativeComponent').default
@@ -221,6 +221,8 @@ function FastImageBase({
     const resolvedSource = Image.resolveAssetSource(
         source as any,
     ) as ImageResolvedAssetSource & { headers: any }
+    // resolvedSource would be frozen, we can't modify it
+    let modifiedSource = resolvedSource
     if (
         resolvedSource?.headers &&
         (FABRIC_ENABLED || Platform.OS === 'android')
@@ -230,7 +232,7 @@ function FastImageBase({
         Object.keys(resolvedSource.headers).forEach((key) => {
             headersArray.push({ name: key, value: resolvedSource.headers[key] })
         })
-        resolvedSource.headers = headersArray
+        modifiedSource = { ...resolvedSource, headers: headersArray }
     }
     const resolvedDefaultSource = resolveDefaultSource(defaultSource)
     const resolvedDefaultSourceAsString =
@@ -242,7 +244,7 @@ function FastImageBase({
                 {...props}
                 tintColor={tintColor}
                 style={StyleSheet.absoluteFill}
-                source={resolvedSource}
+                source={modifiedSource}
                 defaultSource={resolvedDefaultSourceAsString}
                 onFastImageLoadStart={onLoadStart}
                 onFastImageProgress={onProgress}